Add Keys() and Dump()
Keys() returns a string-slice of all the keys in the tree. It does
a depth-first-search of the tree, but using a stack of pointers
instead of recursively calling a node-walking function. The string
slice of keys is in sorted order, naturally.

Dump() is a helper function which println()'s the entire tree.
This is helpful when debugging, to see what the tree looks like.